Add the ability to "stage" generation runs.
This lets you trivially compare what will change between runs of the
tool. While you can copy the resulting "staged" directory structure over
to your production area, that is not recommended. The stage
functionality simply drops deleted assets on the floor with just a log
message to indicate it was deleted. Of course, this game has poor to no
support for deleted assets anyway, so maybe that's a moot point.